Consolidate SportBenefit form tests into unified suite
