+++ docs/sportbenefit_coverage_extensions_2026-05-19.md
@@ -2,10 +2,9 @@
 
 ## Added suites
 
-- Live UI form E2E (staging only): `test_new_web_site_sb/test_live_ui_forms_e2e_sb.py`
-- Post-submit UX (success + backend 4xx/5xx simulation): `test_new_web_site_sb/test_form_post_submit_ux_sb.py`
+- Unified UI forms suite (modal/inline forms, validation, submit endpoints, post-submit UX, live E2E, mobile form checks): `test_new_web_site_sb/test_forms_sb.py`
 - Click-through all key pages (links/buttons/tabs): `test_new_web_site_sb/test_clickthrough_all_pages_sb.py`
-- Mobile functional regression (3 viewports): `test_new_web_site_sb/test_mobile_functional_regression_sb.py`
+- Mobile facilities regression (3 viewports): `test_new_web_site_sb/test_mobile_functional_regression_sb.py`
 - Strict console gate: `test_new_web_site_sb/test_console_strict_sb.py`
 - Legal documents integrity depth checks: `test_new_web_site_sb/test_documents_integrity_sb.py`
 - Visual regression baseline hashes: `test_new_web_site_sb/test_visual_regression_sb.py`
@@ -29,7 +28,7 @@ pytest test_new_web_site_sb -m docs_integrity --headless --b chrome
 pytest test_new_web_site_sb/test_mobile_functional_regression_sb.py --headless --b chrome
 
 # Live UI E2E forms (staging only)
-pytest test_new_web_site_sb/test_live_ui_forms_e2e_sb.py --live-api --base-url https://staging.sportbenefit.eu --headless --b chrome
+pytest test_new_web_site_sb/test_forms_sb.py -m live_api --live-api --base-url https://staging.sportbenefit.eu --headless --b chrome
 
 # Visual regression compare
 pytest test_new_web_site_sb -m visual_regression --headless --b chrome
